add deviceContext to home, accessories, addingDevice, deviceProvider to App
